requirements.txt
flask
pymongo
pycryptodome